+++ game/cli.py
@@ -34,6 +34,8 @@ class Cli:
                 print("\nOpción no válida. Por favor, intenta de nuevo.")
 
     def iniciar_partida(self):
+        # Reiniciar el estado del juego
+        self.__chess__ = Chess()  # Crea un nuevo objeto de Chess para la nueva partida
         print("Partida iniciada...\n")
         print("Las blancas comienzan el juego.\n")
         self.play()  # Llama a la función que maneja la partida
@@ -60,31 +62,44 @@ class Cli:
     def play(self):
         error_message = None  # Variable para almacenar el mensaje de error temporalmente
 
-        while self.__chess__.is_playing():
-            # Mostrar el tablero y el turno actual
+        while self.__chess__.is_playing():  # Mientras el juego no haya terminado
             self.display_board_and_turn()
 
-            # Si hay un error previo, mostrarlo ahora
+            # Mostrar cualquier mensaje de error que se haya producido
             if error_message:
                 print(f'\nError: {error_message}')
                 error_message = None  # Restablecer el mensaje de error
 
-            print("Escribe 'draw' para solicitar un empate, 'menu' para volver al menú o realiza un movimiento.")
+            print("Escribe 'draw' para solicitar un empate, 'rendirse' para rendirte, 'menu' para volver al menú o realiza un movimiento.")
             from_input, to_input = self.get_move_input()
 
             if from_input == 'draw':
                 self.__chess__.request_draw()
                 continue
+            elif from_input == 'rendirse':
+                self.__chess__.rendirse()  # Llamar al método rendirse
+                break  # Terminar el bucle ya que el juego ha finalizado
             elif from_input == 'menu':
                 print("\nVolviendo al menú...")
-                return  # Regresa al menú principal
+                return  # Salir del juego y volver al menú
 
             result = self.attempt_move(from_input, to_input)
             if result:
-                error_message = result  # Almacenar el mensaje de error temporalmente
-            elif self.__chess__.end_game():
+                error_message = result
+            elif self.__chess__.end_game():  # Verificar si el juego ha terminado después de cada movimiento
                 print("\n¡Fin del juego!")
                 break
+
+        # Mensaje de finalización y espera por 'menu'
+        self.wait_for_menu()
+
+    
+    def wait_for_menu(self):
+        while True:
+            option = input("\nEscribe 'menu' para volver al inicio: ").strip().lower()
+            if option == 'menu':
+                print("\nVolviendo al menú principal...")
+                return
 
     def display_board_and_turn(self):
         self.clear_terminal()
@@ -92,25 +107,27 @@ class Cli:
         print(self.__chess__.show_board())  # Muestra el tablero
 
     def get_move_input(self):
-        while True:
-            print('\nIntroduce tu movimiento')
-            from_input = input('Desde (e.g. e2 o draw o menu): ').strip().lower()
-            
-            if from_input == 'draw':
-                return 'draw', None  # Retorna 'draw' si se solicita un empate
-            elif from_input == 'menu':
-                return 'menu', None  # Retorna 'menu' si se solicita volver al menú
+     while True:
+        print('\nIntroduce tu movimiento')
+        from_input = input('Desde (e.g. e2, draw, rendirse o menu): ').strip().lower()
+        
+        if from_input == 'draw':
+            return 'draw', None  # Retorna 'draw' si se solicita un empate
+        elif from_input == 'menu':
+            return 'menu', None  # Retorna 'menu' si se solicita volver al menú
+        elif from_input == 'rendirse':
+            return 'rendirse', None  # Retorna 'rendirse' si se solicita rendirse
 
-            to_input = input('Hasta (e.g. e4): ').strip().lower()
-            
-            try:
-                self.__chess__.parse_position(from_input)  # Verifica formato de entrada
-                self.__chess__.parse_position(to_input)  # Verifica formato de entrada
-                return from_input, to_input
-            except InvalidMove as e:  
-                print(f'\nError en la entrada: {e}\nPor favor, ingresa una posición válida.')
-            except Exception as e:  # Captura errores inesperados
-                print(f'\nError inesperado en la entrada: {e}\nPor favor, intenta de nuevo.')
+        to_input = input('Hasta (e.g. e4): ').strip().lower()
+        
+        try:
+            self.__chess__.parse_position(from_input)  # Verifica formato de entrada
+            self.__chess__.parse_position(to_input)  # Verifica formato de entrada
+            return from_input, to_input
+        except InvalidMove as e:  
+            print(f'\nError en la entrada: {e}\nPor favor, ingresa una posición válida.')
+        except Exception as e:  # Captura errores inesperados
+            print(f'\nError inesperado en la entrada: {e}\nPor favor, intenta de nuevo.')
 
     def attempt_move(self, from_input, to_input):
         error_message = "No se produjo ningún error."
@@ -126,6 +143,8 @@ class Cli:
             error_message = str(e) 
         except OutOfBoard as e:
             error_message = str(e)  
+        except InvalidFormat as e:
+            error_message = str(e)
         except Exception as e:
             error_message = f"Ocurrió un error inesperado: {e}"  # Devuelve el mensaje de error general