+++ game/pieces/knight.py
@@ -12,16 +12,16 @@ class Knight(Piece):
     def mov_correcto(self, from_x, from_y, to_x, to_y):
         # Verifica si el movimiento es a una posición válida
         if not self.valid_positions(from_x, from_y, to_x, to_y):
-            raise InvalidPieceMove("Movimiento inválido a una posición no válida.")
+            raise InvalidPieceMove(piece_name="Caballo")
 
         # Verifica si el movimiento sigue el patrón en forma de L
         if (abs(to_x - from_x), abs(to_y - from_y)) not in [(2, 1), (1, 2)]:
-            raise InvalidPieceMove("Movimiento inválido para el caballo.")
+            raise InvalidPieceMove(piece_name="Caballo")
 
         # Verifica si hay una pieza del mismo color en la posición de destino
         pieza_destino = self.__board__.get_piece(to_x, to_y)
         if pieza_destino and pieza_destino.get_color() == self.get_color():
-            raise InvalidPieceMove("No se puede capturar una pieza propia.")
+            raise InvalidPieceMove("No puedes mover a una posición ocupada por tu propia pieza.")
 
         return True