+++ CHANGELOG.md
@@ -5,9 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+##  [0.3.8] - 2024-10-05
+### Agregando
+- Metodo para rendirce en chess
+
+### Cambios
+- Arreglando test de exceptions
+- Intentando arreglar test en chess
+
 ##  [0.3.7] - 2024-10-04
 ### Cambios
-- Arreglando problema para mostrar mensaje de error en en juego 
+- Arreglando problema para mostrar mensaje de error en el juego 
 
 ##  [0.3.6] - 2024-10-03
 ### Cambios